requirements.txt
# Python-LLM-Labs 根依赖
# 各 lab 若有额外依赖，见其目录下的 requirements.txt

# 数据与校验（lab-07 ~ lab-09）
numpy>=1.24.0
pandas>=2.0.0
pydantic>=2.0.0

# 网络与 API（lab-11 ~ lab-12）
httpx>=0.24.0
requests>=2.28.0
uvicorn[standard]>=0.23.0
fastapi>=0.100.0

# LLM 与 RAG（lab-13 ~ lab-15）
openai>=1.0.0
chromadb>=0.4.0
faiss-cpu>=1.7.4
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-text-splitters>=0.0.1
langchain-chroma>=0.1.0
